refactor/#59 Moved readToken to Utility.java
