+++ src/main/java/Utility.java
@@ -1,6 +1,11 @@
+import java.io.BufferedReader;
 import java.io.File;
+import java.io.FileReader;
+import java.io.IOException;
 
 public class Utility {
+    private static final String tokenFile = ".token";// file path to git access token
+
     /**
      * clearDirectory attempts to clear all files in a given directory.
      * If the program will attempt to delete the file 10 times with a interval on 0.5 seconds.
@@ -30,5 +35,18 @@ public class Utility {
                 }
             }
         }
+    }
+
+    /**
+     * Reads git access token from file
+     *
+     * @return The git access token
+     */
+    public static String readToken() throws IOException {
+        File file = new File(tokenFile);
+        BufferedReader reader = new BufferedReader(new FileReader(file));
+        String token = reader.readLine();
+        reader.close();
+        return token;
     }
 }
